requirements.txt
django==1.6.11
psycopg2==2.6.1
south==1.0.2
django-vanilla-views==1.0.4
django-autoslug==1.9.3
django-colorfield==0.1.4
django-htmlmin==0.8.0
django-uuidfield==0.5.0
python-social-auth==0.2.10
python-twitter==2.2
django-allauth==0.18.0
